Move find_underscores to ide
commit-id:dbebd073
